use old tabgroup layout
